Backend/Prototypes/requirements.txt
awsebcli==3.5.5
backports.ssl-match-hostname==3.4.0.2
blessed==1.9.5
botocore==1.3.7
cement==2.4.0
colorama==0.3.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
Flask==0.10.1
Flask-Login==0.3.2
Flask-MySQL==1.3
Flask-RESTful>=0.3.4
Flask-HTTPAuth>=2.7.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
mysql==0.0.1
MySQL-python==1.2.5
pathspec==0.3.3
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
six==1.10.0
texttable==0.8.4
wcwidth==0.1.5
websocket-client==0.34.0
Werkzeug==0.11
wheel==0.24.0
